requirements.txt
adbutils==1.2.9
apkutils2==1.0.0
asttokens==2.2.1
av==10.0.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
cigam==0.0.3
cmake==3.26.3
comm==0.1.3
contourpy==1.0.7
customtkinter==5.1.3
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
idna==3.4
ipykernel==6.23.1
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.5
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyelftools==0.29
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.30.0
retry==0.9.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.12
tk==0.1.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.6.2
tzdata==2023.3
ultralytics==8.0.111
urllib3==2.0.2
wcwidth==0.2.6
whichcraft==0.6.1
xmltodict==0.13.0
